# GH-115060: Speed up `pathlib.Path.glob()` by not scanning literal parts